% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{Accordion}
\alias{Accordion}
\alias{updateMantineAccordion}
\alias{AccordionItem}
\alias{AccordionControl}
\alias{AccordionPanel}
\title{Mantine Accordion (Shiny stateful input)}
\usage{
Accordion(inputId, ..., value = NULL, multiple = FALSE)

updateMantineAccordion(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)

AccordionItem(...)

AccordionControl(...)

AccordionPanel(...)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a string, or
a vector if \code{multiple = TRUE}) is synced on open/close.}

\item{...}{Other props/children.}

\item{value}{Initial value (an \code{AccordionItem()}'s \code{value}, or a vector
if \code{multiple = TRUE}).}

\item{multiple}{If \code{TRUE}, more than one item can be open at once and
\code{input[[inputId]]} becomes a character vector.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
\code{AccordionItem()} (containing \code{AccordionControl()} + \code{AccordionPanel()})
must be nested inside \code{Accordion()}.
}
