% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buttons.R
\name{ActionIcon}
\alias{ActionIcon}
\title{Mantine ActionIcon (icon-only button)}
\usage{
ActionIcon(...)
}
\arguments{
\item{...}{Props and children (usually a single icon). See
\url{https://mantine.dev/core/action-icon/}.}
}
\description{
Mantine ActionIcon (icon-only button)
}
