% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{ActionIconGroup}
\alias{ActionIconGroup}
\title{Mantine ActionIcon.Group}
\usage{
ActionIconGroup(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/action-icon/#actionicongroup}.}
}
\description{
Groups several \code{\link[=ActionIcon]{ActionIcon()}} visually, the same way \code{\link[=ButtonGroup]{ButtonGroup()}} does
for \code{\link[=Button]{Button()}}.
}
