% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ActionIconGroupSection}
\alias{ActionIconGroupSection}
\title{Mantine ActionIconGroupSection}
\usage{
ActionIconGroupSection(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/action-icon/}.}
}
\description{
Non-interactive section (e.g. a label) placed between grouped ActionIcon()s. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
