% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Alert}
\alias{Alert}
\title{Mantine Alert}
\usage{
Alert(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/alert/}.}
}
\description{
Warning banner with icon/title/color. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
