% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Anchor}
\alias{Anchor}
\title{Mantine Anchor (styled link)}
\usage{
Anchor(...)
}
\arguments{
\item{...}{Props and children (\code{href}, \code{underline}, \code{c}, ...). See
\url{https://mantine.dev/core/anchor/}.}
}
\description{
Mantine Anchor (styled link)
}
