% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{Autocomplete}
\alias{Autocomplete}
\alias{updateMantineAutocomplete}
\title{Mantine Autocomplete (Shiny stateful input)}
\usage{
Autocomplete(inputId, label = NULL, value = "", ...)

updateMantineAutocomplete(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
every keystroke/selection.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props (\code{data}, \code{placeholder}, \code{limit}, ...). See
\url{https://mantine.dev/core/autocomplete/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine Autocomplete (Shiny stateful input)
}
