% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Avatar}
\alias{Avatar}
\title{Mantine Avatar}
\usage{
Avatar(...)
}
\arguments{
\item{...}{Props and children (\code{src}, \code{radius}, \code{color}, \code{size}, ...).
See \url{https://mantine.dev/core/avatar/}.}
}
\description{
Mantine Avatar
}
