% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{AvatarGroup}
\alias{AvatarGroup}
\title{Mantine AvatarGroup}
\usage{
AvatarGroup(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/avatar/}.}
}
\description{
Groups several overlapping Avatars. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
