% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{BackgroundImage}
\alias{BackgroundImage}
\title{Mantine BackgroundImage}
\usage{
BackgroundImage(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/background-image/}.}
}
\description{
Container with a background image. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
