% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{BarsList}
\alias{BarsList}
\title{Mantine BarsList (list of horizontal bars with names and values)}
\usage{
BarsList(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} (or list of rows) with \code{name} and \code{value}
columns.}

\item{...}{Other props (\code{barsLabel}, \code{valueLabel}, \code{barGap},
\code{minBarSize}, \code{barHeight}, ...). See
\url{https://mantine.dev/charts/bars-list/}.}
}
\description{
Mantine BarsList (list of horizontal bars with names and values)
}
\examples{
\dontrun{
BarsList(data = data.frame(
  name = c("React", "Vue", "Angular"),
  value = c(950000, 320000, 580000)
))
}
}
