% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Blockquote}
\alias{Blockquote}
\title{Mantine Blockquote}
\usage{
Blockquote(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/blockquote/}.}
}
\description{
Styled quotation block. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
