% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Box}
\alias{Box}
\title{Mantine Box (generic styleable container)}
\usage{
Box(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/box/}.}
}
\description{
Mantine Box (generic styleable container)
}
