% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Breadcrumbs}
\alias{Breadcrumbs}
\title{Mantine Breadcrumbs}
\usage{
Breadcrumbs(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/breadcrumbs/}.}
}
\description{
Navigation trail (Home > Section > Page). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
