% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Burger}
\alias{Burger}
\title{Mantine Burger (hamburger icon for a responsive navbar)}
\usage{
Burger(...)
}
\arguments{
\item{...}{Props and children. To wire it to Shiny use \code{\link[=navbarBurger]{navbarBurger()}}.}
}
\description{
Mantine Burger (hamburger icon for a responsive navbar)
}
