% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buttons.R
\name{ButtonGroup}
\alias{ButtonGroup}
\title{Mantine Button.Group (visually groups several Button together)}
\usage{
ButtonGroup(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/button/#buttongroup}.}
}
\description{
Mantine Button.Group (visually groups several Button together)
}
