% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buttons.R
\name{ButtonWithMenu}
\alias{ButtonWithMenu}
\title{Button with menu}
\usage{
ButtonWithMenu(label, ..., color = "blue")
}
\arguments{
\item{label}{Button label.}

\item{...}{Menu items (typically \code{\link[=menuItem]{menuItem()}}) and other props
forwarded to the \code{Button} (\code{variant}, \code{size}, ...).}

\item{color}{Button color.}
}
\description{
A button that opens a dropdown menu on click. Items should be passed as
\code{\link[=menuItem]{menuItem()}} (or \code{\link[=MenuLabel]{MenuLabel()}}/\code{\link[=MenuDivider]{MenuDivider()}}) among the \code{...}.
}
\examples{
\dontrun{
ButtonWithMenu(
  "Create new",
  menuItem("action", "project", "Project"),
  menuItem("action", "folder", "Folder")
)
}
}
