% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Card}
\alias{Card}
\title{Mantine Card}
\usage{
Card(...)
}
\arguments{
\item{...}{Props and children forwarded to Mantine's \code{Card} component
(\code{shadow}, \code{padding}, \code{radius}, \code{withBorder}, ...). See
\url{https://mantine.dev/core/card/}.}
}
\description{
Mantine Card
}
