% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carousel.R
\name{Carousel}
\alias{Carousel}
\alias{CarouselSlide}
\title{Mantine Carousel}
\usage{
Carousel(...)

CarouselSlide(...)
}
\arguments{
\item{...}{Props and children (\code{\link[=CarouselSlide]{CarouselSlide()}}) for \code{Carousel()};
props for \code{CarouselSlide()}. See \url{https://mantine.dev/x/carousel/}.}
}
\description{
Requires \verb{@mantine/carousel} + \code{embla-carousel-react} (already included
in the package's JS bundle).
}
