% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Center}
\alias{Center}
\title{Mantine Center (centers content horizontally and vertically)}
\usage{
Center(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/center/}.}
}
\description{
Mantine Center (centers content horizontally and vertically)
}
