% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{Checkbox}
\alias{Checkbox}
\alias{updateMantineCheckbox}
\title{Mantine Checkbox (Shiny stateful boolean input)}
\usage{
Checkbox(inputId, label = NULL, value = FALSE, ...)

updateMantineCheckbox(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (\code{TRUE}/\code{FALSE})
is synced on every toggle.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props (\code{description}, \code{color}, \code{size}, ...). See
\url{https://mantine.dev/core/checkbox/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine Checkbox (Shiny stateful boolean input)
}
