% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{CheckIcon}
\alias{CheckIcon}
\title{Mantine CheckIcon}
\usage{
CheckIcon(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/checkbox/}.}
}
\description{
The checkmark icon Checkbox()/CheckboxCard() render when checked. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
