% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modals.R
\name{closeAllMantineModals}
\alias{closeAllMantineModals}
\title{Close all imperatively-opened modals (\verb{@mantine/modals})}
\usage{
closeAllMantineModals(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Session object passed to the Shiny server function.}
}
\description{
Close all imperatively-opened modals (\verb{@mantine/modals})
}
