% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{CloseButton}
\alias{CloseButton}
\title{Mantine CloseButton}
\usage{
CloseButton(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/close-button/}.}
}
\description{
"x" icon button for closing/removing. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
