% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{CloseIcon}
\alias{CloseIcon}
\title{Mantine CloseIcon}
\usage{
CloseIcon(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/close-button/}.}
}
\description{
The "x" icon CloseButton() renders. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
