% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modals.R
\name{closeMantineModal}
\alias{closeMantineModal}
\title{Close an imperatively-opened modal (\verb{@mantine/modals})}
\usage{
closeMantineModal(session = shiny::getDefaultReactiveDomain(), id)
}
\arguments{
\item{session}{Session object passed to the Shiny server function.}

\item{id}{Id of the modal (the one returned/assigned when it was opened).}
}
\description{
Close an imperatively-opened modal (\verb{@mantine/modals})
}
