% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Code}
\alias{Code}
\title{Mantine Code}
\usage{
Code(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/code/}.}
}
\description{
Inline or block code-styled text. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
