% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHighlight.R
\name{CodeHighlightTabs}
\alias{CodeHighlightTabs}
\title{Mantine CodeHighlightTabs (tabbed multi-file code block)}
\usage{
CodeHighlightTabs(code, ...)
}
\arguments{
\item{code}{A list of \code{list(code = ..., language = ..., fileName = ...)},
one per tab.}

\item{...}{Other props (\code{withCopyButton}, \code{defaultActiveTab}, ...). See
\url{https://mantine.dev/x/code-highlight/#codehighlighttabs}.}
}
\description{
Like \code{\link[=CodeHighlight]{CodeHighlight()}} (same plain-text adapter, no syntax coloring), but
displays several code snippets as a tabbed file browser.
}
\examples{
\dontrun{
CodeHighlightTabs(code = list(
  list(fileName = "app.R", language = "r", code = "shiny::runApp()"),
  list(fileName = "index.js", language = "js", code = "console.log('hi')")
))
}
}
