% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{ColorInput}
\alias{ColorInput}
\alias{updateMantineColorInput}
\title{Mantine ColorInput (Shiny stateful input)}
\usage{
ColorInput(inputId, label = NULL, value = "#000000", ...)

updateMantineColorInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a color
string) is synced to the selection.}

\item{label}{Field label.}

\item{value}{Initial value (e.g. \code{"#228BE6"}).}

\item{...}{Other props (\code{format}, \code{swatches}, ...). See
\url{https://mantine.dev/core/color-input/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine ColorInput (Shiny stateful input)
}
