% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{ColorPicker}
\alias{ColorPicker}
\alias{updateMantineColorPicker}
\title{Mantine ColorPicker (Shiny stateful input, inline picker)}
\usage{
ColorPicker(inputId, value = "#000000", ...)

updateMantineColorPicker(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced to
the selection.}

\item{value}{Initial value.}

\item{...}{Other props (\code{format}, \code{swatches}, ...). See
\url{https://mantine.dev/core/color-picker/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine ColorPicker (Shiny stateful input, inline picker)
}
