% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ColorSwatch}
\alias{ColorSwatch}
\title{Mantine ColorSwatch}
\usage{
ColorSwatch(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/color-swatch/}.}
}
\description{
Small color sample. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
