% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Container}
\alias{Container}
\title{Mantine Container}
\usage{
Container(...)
}
\arguments{
\item{...}{Props and children (\code{size}, \code{fluid}, ...). See
\url{https://mantine.dev/core/container/}.}
}
\description{
Mantine Container
}
