% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{DataList}
\alias{DataList}
\alias{DataListItem}
\alias{DataListItemLabel}
\alias{DataListItemValue}
\title{Mantine DataList}
\usage{
DataList(...)

DataListItem(...)

DataListItemLabel(...)

DataListItemValue(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/data-list/}.}
}
\description{
Key/value list (definition-list style). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
