% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{DatePicker}
\alias{DatePicker}
\alias{updateMantineDatePicker}
\title{Mantine DatePicker (always-visible inline calendar)}
\usage{
DatePicker(inputId, value = NULL, ...)

updateMantineDatePicker(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a
\code{"YYYY-MM-DD"} string) is synced on selection.}

\item{value}{Initial value.}

\item{...}{Other props (\code{type}, \code{minDate}, \code{maxDate}, ...). See
\url{https://mantine.dev/dates/date-picker/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine DatePicker (always-visible inline calendar)
}
