% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{DatesProvider}
\alias{DatesProvider}
\title{Mantine DatesProvider (locale/date format for its children)}
\usage{
DatesProvider(...)
}
\arguments{
\item{...}{Props (\code{locale}, \code{firstDayOfWeek}, ...) and children. See
\url{https://mantine.dev/dates/dates-provider/}.}
}
\description{
Mantine DatesProvider (locale/date format for its children)
}
