% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{DateTimePicker}
\alias{DateTimePicker}
\alias{updateMantineDateTimePicker}
\title{Mantine DateTimePicker (Shiny stateful input, date + time)}
\usage{
DateTimePicker(inputId, label = NULL, value = NULL, ...)

updateMantineDateTimePicker(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a
\code{"YYYY-MM-DD HH:mm:ss"} string) is synced on selection.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props. See \url{https://mantine.dev/dates/date-time-picker/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine DateTimePicker (Shiny stateful input, date + time)
}
