% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Divider}
\alias{Divider}
\title{Mantine Divider (separator line)}
\usage{
Divider(...)
}
\arguments{
\item{...}{Props (\code{label}, \code{my}, ...). See \url{https://mantine.dev/core/divider/}.}
}
\description{
Mantine Divider (separator line)
}
