% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{DrawerBody}
\alias{DrawerBody}
\title{Mantine DrawerBody}
\usage{
DrawerBody(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/drawer/}.}
}
\description{
Body/content area of a DrawerRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
