% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{DrawerContent}
\alias{DrawerContent}
\title{Mantine DrawerContent}
\usage{
DrawerContent(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/drawer/}.}
}
\description{
Content box of a DrawerRoot() (holds DrawerHeader()/DrawerBody()). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
