% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{DrawerHeader}
\alias{DrawerHeader}
\title{Mantine DrawerHeader}
\usage{
DrawerHeader(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/drawer/}.}
}
\description{
Header row of a DrawerRoot() (usually DrawerTitle() + DrawerCloseButton()). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
