% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{DrawerTitle}
\alias{DrawerTitle}
\title{Mantine DrawerTitle}
\usage{
DrawerTitle(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/drawer/}.}
}
\description{
Title text of a DrawerRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
