% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{EmptyState}
\alias{EmptyState}
\alias{EmptyStateActions}
\alias{EmptyStateDescription}
\alias{EmptyStateIndicator}
\alias{EmptyStateTitle}
\title{Mantine EmptyState}
\usage{
EmptyState(...)

EmptyStateActions(...)

EmptyStateDescription(...)

EmptyStateIndicator(...)

EmptyStateTitle(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/empty-state/}.}
}
\description{
Placeholder shown when there is no content to display. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
