% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Fieldset}
\alias{Fieldset}
\title{Mantine Fieldset}
\usage{
Fieldset(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/fieldset/}.}
}
\description{
Groups form inputs under a legend. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
