% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Flex}
\alias{Flex}
\title{Mantine Flex}
\usage{
Flex(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/flex/}.}
}
\description{
Flexbox container with shorthand props (direction, gap, wrap, align, justify). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
