% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Grid}
\alias{Grid}
\alias{GridCol}
\title{Mantine Grid family (12-column grid)}
\usage{
Grid(...)

GridCol(...)
}
\arguments{
\item{...}{Props and children (\code{gutter}, ...) for \code{Grid()}; (\code{span},
\code{offset}, ...) for \code{GridCol()}. See \url{https://mantine.dev/core/grid/}.}
}
\description{
\code{GridCol()} must be nested inside \code{Grid()}.
}
