% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Group}
\alias{Group}
\title{Mantine Group (horizontal flex layout)}
\usage{
Group(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/group/}.}
}
\description{
Mantine Group (horizontal flex layout)
}
