% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buttons.R, R/components.R
\name{IconSun}
\alias{IconSun}
\alias{IconMoon}
\alias{IconCopy}
\alias{IconCheck}
\alias{IconBrandGoogle}
\alias{IconBrandFacebook}
\alias{IconBrandGithub}
\alias{IconBrandDiscord}
\alias{IconHome2}
\alias{IconLayoutDashboard}
\alias{IconSettings}
\alias{IconUsers}
\alias{IconPlus}
\alias{IconBrandTwitter}
\alias{IconBrandInstagram}
\alias{IconBrandYoutube}
\alias{IconBrandLinkedin}
\alias{IconArrowUpRight}
\alias{IconArrowDownRight}
\alias{IconTrendingUp}
\alias{IconTrendingDown}
\alias{IconPhone}
\alias{IconMail}
\alias{IconGripVertical}
\alias{IconUpload}
\alias{IconX}
\alias{IconMapPin}
\alias{IconHeart}
\alias{IconStar}
\alias{IconDots}
\alias{IconEdit}
\alias{IconTrash}
\alias{IconLogout}
\alias{IconChevronRight}
\alias{IconSearch}
\alias{IconChevronDown}
\alias{IconCode}
\alias{IconBook}
\alias{IconChartBar}
\alias{IconFingerprint}
\alias{IconChartLine}
\alias{IconChartPie}
\alias{IconFileText}
\alias{IconBell}
\alias{IconExternalLink}
\alias{IconDownload}
\alias{IconCurrencyDollar}
\alias{IconPercentage}
\alias{IconCategory}
\title{Tabler icons}
\usage{
IconSun(...)

IconMoon(...)

IconCopy(...)

IconCheck(...)

IconBrandGoogle(...)

IconBrandFacebook(...)

IconBrandGithub(...)

IconBrandDiscord(...)

IconHome2(...)

IconLayoutDashboard(...)

IconSettings(...)

IconUsers(...)

IconPlus(...)

IconBrandTwitter(...)

IconBrandInstagram(...)

IconBrandYoutube(...)

IconBrandLinkedin(...)

IconArrowUpRight(...)

IconArrowDownRight(...)

IconTrendingUp(...)

IconTrendingDown(...)

IconPhone(...)

IconMail(...)

IconGripVertical(...)

IconUpload(...)

IconX(...)

IconMapPin(...)

IconHeart(...)

IconStar(...)

IconDots(...)

IconEdit(...)

IconTrash(...)

IconLogout(...)

IconChevronRight(...)

IconSearch(...)

IconChevronDown(...)

IconCode(...)

IconBook(...)

IconChartBar(...)

IconFingerprint(...)

IconChartLine(...)

IconChartPie(...)

IconFileText(...)

IconBell(...)

IconExternalLink(...)

IconDownload(...)

IconCurrencyDollar(...)

IconPercentage(...)

IconCategory(...)
}
\arguments{
\item{...}{Props (\code{size}, \code{color}, \code{stroke}, ...).}
}
\description{
A small subset of \url{https://tabler.io/icons} already registered in the
JS bundle. Common props: \code{size}, \code{color}, \code{stroke}.
}
