% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{Image}
\alias{Image}
\title{Mantine Image}
\usage{
Image(...)
}
\arguments{
\item{...}{Props (\code{src}, \code{radius}, \code{fit}, \code{h}, ...). See
\url{https://mantine.dev/core/image/}.}
}
\description{
Mantine Image
}
