% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{InputBase}
\alias{InputBase}
\title{Mantine InputBase}
\usage{
InputBase(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/input/}.}
}
\description{
Like Input(), extended with a few more style-composition options used internally by other inputs. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
