% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{InputClearButton}
\alias{InputClearButton}
\title{Mantine InputClearButton}
\usage{
InputClearButton(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/input/}.}
}
\description{
The "x" clear button Select()/DatePickerInput()/etc. render when clearable, exposed for custom inputs built on Input(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
