% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{InputWrapper}
\alias{InputWrapper}
\title{Mantine InputWrapper}
\usage{
InputWrapper(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/input/}.}
}
\description{
Label/description/error/required-asterisk chrome around arbitrary custom content. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
