% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Kbd}
\alias{Kbd}
\title{Mantine Kbd}
\usage{
Kbd(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/kbd/}.}
}
\description{
Keyboard-key-styled text. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
