% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{List}
\alias{List}
\alias{ListItem}
\title{Mantine List}
\usage{
List(...)

ListItem(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/list/}.}
}
\description{
Bulleted/numbered list. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
