% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Loader}
\alias{Loader}
\title{Mantine Loader}
\usage{
Loader(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/loader/}.}
}
\description{
Loading indicator (spinner). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
