% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlays.R
\name{LoadingOverlay}
\alias{LoadingOverlay}
\title{Mantine LoadingOverlay}
\usage{
LoadingOverlay(visible = FALSE, ...)
}
\arguments{
\item{visible}{If \code{TRUE}, shows the loading overlay. Pass \code{mantineId}
among the \code{...} to be able to toggle it from R with
\code{\link[=updateMantineProps]{updateMantineProps()}}.}

\item{...}{Other props (\code{overlayProps}, \code{loaderProps}, ...). See
\url{https://mantine.dev/core/loading-overlay/}.}
}
\description{
Mantine LoadingOverlay
}
