% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{mantineDependency}
\alias{mantineDependency}
\title{Mantine JS/CSS dependency}
\usage{
mantineDependency()
}
\value{
An \code{htmlDependency} object.
}
\description{
HTML dependency that loads the compiled 'shiny.mantine' JavaScript bundle
(the 'Mantine' UI components together with its own bundled copy of
'React' 19) and its associated styles.
}
