% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Mark}
\alias{Mark}
\title{Mantine Mark}
\usage{
Mark(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/mark/}.}
}
\description{
Highlighted text (\verb{<mark>} tag). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
