% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Marquee}
\alias{Marquee}
\title{Mantine Marquee}
\usage{
Marquee(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/marquee/}.}
}
\description{
Horizontally scrolling ticker of content. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
