% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buttons.R
\name{Menu}
\alias{Menu}
\alias{MenuTarget}
\alias{MenuDropdown}
\alias{MenuLabel}
\alias{MenuDivider}
\title{Mantine Menu family (dropdown on click)}
\usage{
Menu(...)

MenuTarget(...)

MenuDropdown(...)

MenuLabel(...)

MenuDivider(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/menu/}.}
}
\description{
Generic base for dropdown menus. For the ready-to-use "Button with menu"
and "Split button" patterns see \code{\link[=ButtonWithMenu]{ButtonWithMenu()}} and \code{\link[=SplitButton]{SplitButton()}};
use these primitives only if you need a \code{Menu} with a target other than
a button.
}
