% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{Menubar}
\alias{Menubar}
\alias{MenubarMenu}
\alias{MenubarTarget}
\alias{MenubarDropdown}
\title{Mantine Menubar}
\usage{
Menubar(...)

MenubarMenu(...)

MenubarTarget(...)

MenubarDropdown(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/menubar/}.}
}
\description{
A horizontal bar of menus (File/Edit/View-style), distinct from a single
dropdown \code{\link[=Menu]{Menu()}}. \code{MenubarMenu()} groups one menu's target + dropdown;
items inside \code{MenubarDropdown()} are plain \code{\link[=menuItem]{menuItem()}} calls.
}
