% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{MenuSearch}
\alias{MenuSearch}
\alias{MenuContextMenu}
\title{Mantine Menu.Search / Menu.ContextMenu}
\usage{
MenuSearch(...)

MenuContextMenu(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/menu/}.}
}
\description{
\code{MenuSearch()} renders a search field inside a \code{\link[=MenuDropdown]{MenuDropdown()}}. Wrap a
target in \code{MenuContextMenu()} instead of \code{\link[=MenuTarget]{MenuTarget()}} to open the menu
on right-click instead of left-click.
}
