% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{MenuSub}
\alias{MenuSub}
\alias{MenuSubTarget}
\alias{MenuSubDropdown}
\alias{menuSubItem}
\title{Mantine Menu submenu (\code{Menu.Sub})}
\usage{
MenuSub(...)

MenuSubTarget(...)

MenuSubDropdown(...)

menuSubItem(inputId, value, label, ...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/menu/#nested-menus}.}

\item{inputId}{Id of the Shiny input that receives the selected value.}

\item{value}{Value sent to Shiny when the item is clicked.}

\item{label}{Item text.}
}
\description{
Nests a submenu inside a \code{\link[=MenuDropdown]{MenuDropdown()}} item. \code{MenuSubItem()} behaves
like \code{\link[=menuItem]{menuItem()}} (reports \code{value} to \code{input[[inputId]]} on click).
}
