% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{MiniCalendar}
\alias{MiniCalendar}
\alias{updateMantineMiniCalendar}
\title{Mantine MiniCalendar (compact single-week calendar strip)}
\usage{
MiniCalendar(inputId, value = NULL, ...)

updateMantineMiniCalendar(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a
\code{"YYYY-MM-DD"} string) is synced on selection.}

\item{value}{Initial value.}

\item{...}{Other props. See \url{https://mantine.dev/dates/mini-calendar/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine MiniCalendar (compact single-week calendar strip)
}
