% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ModalBody}
\alias{ModalBody}
\title{Mantine ModalBody}
\usage{
ModalBody(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/modal/}.}
}
\description{
Body/content area of a ModalRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
