% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ModalCloseButton}
\alias{ModalCloseButton}
\title{Mantine ModalCloseButton}
\usage{
ModalCloseButton(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/modal/}.}
}
\description{
Close ("x") button of a ModalRoot() - already wired to its onClose. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
