% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ModalContent}
\alias{ModalContent}
\title{Mantine ModalContent}
\usage{
ModalContent(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/modal/}.}
}
\description{
Content box of a ModalRoot() (holds ModalHeader()/ModalBody()). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
