% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ModalHeader}
\alias{ModalHeader}
\title{Mantine ModalHeader}
\usage{
ModalHeader(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/modal/}.}
}
\description{
Header row of a ModalRoot() (usually ModalTitle() + ModalCloseButton()). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
