% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ModalOverlay}
\alias{ModalOverlay}
\title{Mantine ModalOverlay}
\usage{
ModalOverlay(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/modal/}.}
}
\description{
Dimmed backdrop behind a ModalRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
