% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{ModalTitle}
\alias{ModalTitle}
\title{Mantine ModalTitle}
\usage{
ModalTitle(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/modal/}.}
}
\description{
Title text of a ModalRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
