% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{MonthPickerInput}
\alias{MonthPickerInput}
\alias{updateMantineMonthPickerInput}
\title{Mantine MonthPickerInput (Shiny stateful input, month/year selection)}
\usage{
MonthPickerInput(inputId, label = NULL, value = NULL, ...)

updateMantineMonthPickerInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a
\code{"YYYY-MM-DD"} string, first day of the month) is synced on selection.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props. See \url{https://mantine.dev/dates/month-picker-input/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine MonthPickerInput (Shiny stateful input, month/year selection)
}
