% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{NativeSelect}
\alias{NativeSelect}
\alias{updateMantineNativeSelect}
\title{Mantine NativeSelect (Shiny stateful input, native \verb{<select>})}
\usage{
NativeSelect(inputId, label = NULL, value = NULL, ...)

updateMantineNativeSelect(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced to
the selection.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props (\code{data}, ...). See
\url{https://mantine.dev/core/native-select/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine NativeSelect (Shiny stateful input, native \verb{<select>})
}
