% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{NavLink}
\alias{NavLink}
\title{Mantine NavLink}
\usage{
NavLink(...)
}
\arguments{
\item{...}{Props and children forwarded to Mantine's \code{NavLink}
component. For an item wired to Shiny use \code{\link[=navLinkItem]{navLinkItem()}}.}
}
\description{
Mantine NavLink
}
