% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{NumberInput}
\alias{NumberInput}
\alias{updateMantineNumberInput}
\title{Mantine NumberInput (Shiny stateful input)}
\usage{
NumberInput(inputId, label = NULL, value = NULL, ...)

updateMantineNumberInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
every change.}

\item{label}{Field label.}

\item{value}{Initial value (a number or \code{NULL}).}

\item{...}{Other props (\code{min}, \code{max}, \code{step}, \code{prefix}, \code{suffix}, ...).
See \url{https://mantine.dev/core/number-input/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine NumberInput (Shiny stateful input)
}
