% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modals.R
\name{openMantineConfirmModal}
\alias{openMantineConfirmModal}
\title{Open a confirmation modal (\verb{@mantine/modals})}
\usage{
openMantineConfirmModal(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{session}{Session object passed to the Shiny server function.}

\item{inputId}{Id of the Shiny input that receives \code{TRUE}/\code{FALSE}.}

\item{...}{Other props (\code{title}, \code{message}, \code{labels}, \code{confirmProps},
\code{cancelProps}, ...).}
}
\description{
On confirm, \code{input[[inputId]]} receives \code{TRUE}; on cancel, \code{FALSE}.
}
