% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modals.R
\name{openMantineModal}
\alias{openMantineModal}
\title{Open a generic modal (\verb{@mantine/modals})}
\usage{
openMantineModal(session = shiny::getDefaultReactiveDomain(), ...)
}
\arguments{
\item{session}{Session object passed to the Shiny server function.}

\item{...}{Props (\code{title}, \code{children} as plain text, \code{size}, ...).}
}
\description{
Open a generic modal (\verb{@mantine/modals})
}
