% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Overlay}
\alias{Overlay}
\title{Mantine Overlay}
\usage{
Overlay(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/overlay/}.}
}
\description{
Semi-transparent layer over some content. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
