% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{Pagination}
\alias{Pagination}
\alias{updateMantinePagination}
\title{Mantine Pagination (Shiny stateful input)}
\usage{
Pagination(inputId, total, value = 1, ...)

updateMantinePagination(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
page change.}

\item{total}{Total number of pages.}

\item{value}{Initial page.}

\item{...}{Other props (\code{siblings}, \code{boundaries}, ...). See
\url{https://mantine.dev/core/pagination/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine Pagination (Shiny stateful input)
}
