% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{PaginationControl}
\alias{PaginationControl}
\title{Mantine PaginationControl}
\usage{
PaginationControl(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/pagination/}.}
}
\description{
Generic clickable pagination-styled button, for building entirely custom page controls inside PaginationRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
