% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{PaginationDots}
\alias{PaginationDots}
\title{Mantine PaginationDots}
\usage{
PaginationDots(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/pagination/}.}
}
\description{
Ellipsis ("...") separator of a PaginationRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
