% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{PaginationItems}
\alias{PaginationItems}
\title{Mantine PaginationItems}
\usage{
PaginationItems(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/pagination/}.}
}
\description{
Numbered page buttons of a PaginationRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
