% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{PaginationNext}
\alias{PaginationNext}
\title{Mantine PaginationNext}
\usage{
PaginationNext(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/pagination/}.}
}
\description{
"Go to next page" button of a PaginationRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
