% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{PaginationPrevious}
\alias{PaginationPrevious}
\title{Mantine PaginationPrevious}
\usage{
PaginationPrevious(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/pagination/}.}
}
\description{
"Go to previous page" button of a PaginationRoot(). Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
