% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{PasswordInput}
\alias{PasswordInput}
\alias{updateMantinePasswordInput}
\title{Mantine PasswordInput (Shiny stateful input)}
\usage{
PasswordInput(inputId, label = NULL, value = "", ...)

updateMantinePasswordInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
every keystroke.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props (\code{placeholder}, \code{description}, ...). See
\url{https://mantine.dev/core/password-input/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine PasswordInput (Shiny stateful input)
}
