% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{PieChart}
\alias{PieChart}
\alias{DonutChart}
\title{Mantine PieChart / DonutChart (pie/donut charts)}
\usage{
PieChart(data, ...)

DonutChart(data, ...)
}
\arguments{
\item{data}{Chart data (\code{data.frame} or list of rows, with
\code{name}/\code{value}/\code{color}).}

\item{...}{Other props (\code{withLabels}, \code{withTooltip}, \code{size}, ...). See
\url{https://mantine.dev/charts/pie-chart/}.}
}
\description{
\code{data} is a \code{data.frame} (or list of rows) with columns \code{name}, \code{value},
and optionally \code{color}.
}
\examples{
\dontrun{
PieChart(data = data.frame(
  name = c("A", "B", "C"),
  value = c(40, 35, 25),
  color = c("blue.6", "orange.6", "grape.6")
))
}
}
