% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Pill}
\alias{Pill}
\alias{PillGroup}
\title{Mantine Pill}
\usage{
Pill(...)

PillGroup(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/pill/}.}
}
\description{
Small rounded tag/chip element. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
