% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{PillsInput}
\alias{PillsInput}
\alias{PillsInputField}
\title{Mantine PillsInput}
\usage{
PillsInput(...)

PillsInputField(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/pills-input/}.}
}
\description{
Styled input box for a list of Pill()s plus a text field, the multi-value building block MultiSelect()/TagsInput() use internally. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
