% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{PinInput}
\alias{PinInput}
\alias{updateMantinePinInput}
\title{Mantine PinInput (Shiny stateful input, e.g. an OTP code)}
\usage{
PinInput(inputId, value = "", ...)

updateMantinePinInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
every digit entered.}

\item{value}{Initial value.}

\item{...}{Other props (\code{length}, \code{type}, \code{mask}, ...). See
\url{https://mantine.dev/core/pin-input/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine PinInput (Shiny stateful input, e.g. an OTP code)
}
