% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{Portal}
\alias{Portal}
\title{Mantine Portal (render children into a different DOM node)}
\usage{
Portal(...)
}
\arguments{
\item{...}{Props and children (\code{target}, \code{reuseTargetNode}, ...). See
\url{https://mantine.dev/core/portal/}.}
}
\description{
Mantine Portal (render children into a different DOM node)
}
