% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buttons.R
\name{Progress}
\alias{Progress}
\title{Mantine Progress (progress bar)}
\usage{
Progress(...)
}
\arguments{
\item{...}{Props (\code{value}, \code{color}, \code{size}, ...). See
\url{https://mantine.dev/core/progress/}.}
}
\description{
Mantine Progress (progress bar)
}
