% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{RadarChart}
\alias{RadarChart}
\alias{CompositeChart}
\title{Mantine RadarChart / CompositeChart (multi-series charts)}
\usage{
RadarChart(data, series, dataKey, ...)

CompositeChart(data, series, ...)
}
\arguments{
\item{data}{Chart data (\code{data.frame} or list of rows).}

\item{series}{List of \code{list(name = ..., color = ...)} (\code{CompositeChart()}
series can also set \code{type = "line"/"bar"/"area"}).}

\item{dataKey}{Column used for radar axis labels.}

\item{...}{Other props. See \url{https://mantine.dev/charts/radar-chart/} /
\url{https://mantine.dev/charts/composite-chart/}.}
}
\description{
\code{RadarChart()} plots one or more series across radial axes (\code{dataKey}
picks the column used for axis labels). \code{CompositeChart()} is like
\code{LineChart()}/\code{BarChart()}/\code{AreaChart()} but each series in \code{series} can
set its own \code{type} (\code{"line"}/\code{"bar"}/\code{"area"}), combining them in one
chart. Both accept \code{data} as a \code{data.frame} or list of rows, same as
\code{\link[=LineChart]{LineChart()}}.
}
