% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{RadioIcon}
\alias{RadioIcon}
\title{Mantine RadioIcon}
\usage{
RadioIcon(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/radio/}.}
}
\description{
The dot icon Radio()/RadioCard() render when checked. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
