% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sliders.R
\name{RangeSlider}
\alias{RangeSlider}
\alias{updateMantineRangeSlider}
\title{Mantine RangeSlider (Shiny stateful input, range)}
\usage{
RangeSlider(inputId, value = c(20, 80), ...)

updateMantineRangeSlider(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a vector of 2
numbers) is synced by dragging the handles.}

\item{value}{Initial value, a vector of 2 numbers (e.g. \code{c(20, 80)}).}

\item{...}{Other props (\code{min}, \code{max}, \code{step}, \code{marks}, ...). See
\url{https://mantine.dev/core/slider/#rangeslider}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine RangeSlider (Shiny stateful input, range)
}
