% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{Rating}
\alias{Rating}
\alias{updateMantineRating}
\title{Mantine Rating (Shiny stateful input)}
\usage{
Rating(inputId, value = 0, ...)

updateMantineRating(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced to
the selection.}

\item{value}{Initial value.}

\item{...}{Other props (\code{count}, \code{fractions}, \code{color}, ...). See
\url{https://mantine.dev/core/rating/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine Rating (Shiny stateful input)
}
