% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{RemoveScroll}
\alias{RemoveScroll}
\title{Mantine RemoveScroll}
\usage{
RemoveScroll(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/floating/}.}
}
\description{
Prevents the page behind it from scrolling while enabled. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
