% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{RingProgress}
\alias{RingProgress}
\title{Mantine RingProgress}
\usage{
RingProgress(...)
}
\arguments{
\item{...}{Props (\code{sections}, \code{label}, \code{size}, \code{thickness}, ...). See
\url{https://mantine.dev/core/ring-progress/}.}
}
\description{
Mantine RingProgress
}
