% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{RollingNumber}
\alias{RollingNumber}
\title{Mantine RollingNumber}
\usage{
RollingNumber(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/rolling-number/}.}
}
\description{
Animated rolling/odometer-style number display. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
