% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ScrollArea}
\alias{ScrollArea}
\title{Mantine ScrollArea}
\usage{
ScrollArea(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/scroll-area/}.}
}
\description{
Mantine ScrollArea
}
