% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{SegmentedControl}
\alias{SegmentedControl}
\alias{updateMantineSegmentedControl}
\title{Mantine SegmentedControl (Shiny stateful input)}
\usage{
SegmentedControl(inputId, data, value = NULL, ...)

updateMantineSegmentedControl(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
every selection.}

\item{data}{Vector or list of options (\code{value}/\code{label}). See
\url{https://mantine.dev/core/segmented-control/}.}

\item{value}{Initial value.}

\item{...}{Other props (\code{color}, \code{fullWidth}, ...).}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine SegmentedControl (Shiny stateful input)
}
