% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Select}
\alias{Select}
\alias{updateMantineSelect}
\title{Mantine Select (Shiny stateful input)}
\usage{
Select(inputId, label = NULL, value = NULL, ...)

updateMantineSelect(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
every selection.}

\item{label}{Field label.}

\item{value}{Initial value (or \code{NULL}).}

\item{...}{Other props (\code{data}, \code{placeholder}, \code{searchable}, ...). See
\url{https://mantine.dev/core/select/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine Select (Shiny stateful input)
}
