% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{SemiCircleProgress}
\alias{SemiCircleProgress}
\title{Mantine SemiCircleProgress}
\usage{
SemiCircleProgress(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/semi-circle-progress/}.}
}
\description{
Half-circle progress indicator. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
