% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notifications.R
\name{showMantineNotification}
\alias{showMantineNotification}
\alias{hideMantineNotification}
\title{Show/hide a Mantine notification}
\usage{
showMantineNotification(session = shiny::getDefaultReactiveDomain(), ...)

hideMantineNotification(session = shiny::getDefaultReactiveDomain(), id)
}
\arguments{
\item{session}{Session object passed to the Shiny server function.}

\item{...}{Notification props (\code{title}, \code{message}, \code{color}, \code{icon},
\code{autoClose}, \code{withCloseButton}, \code{id}, ...). See
\url{https://mantine.dev/x/notifications/#notifications-system}.}

\item{id}{Id of the notification to hide (the one passed to
\code{showMantineNotification(id = ...)}).}
}
\value{
None. Called for its side effect.
}
\description{
Show/hide a Mantine notification
}
