% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{SimpleGrid}
\alias{SimpleGrid}
\title{Mantine SimpleGrid (responsive grid with uniform columns)}
\usage{
SimpleGrid(...)
}
\arguments{
\item{...}{Props and children (\code{cols}, \code{spacing}, ...). See
\url{https://mantine.dev/core/simple-grid/}.}
}
\description{
Mantine SimpleGrid (responsive grid with uniform columns)
}
