% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Skeleton}
\alias{Skeleton}
\title{Mantine Skeleton}
\usage{
Skeleton(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/skeleton/}.}
}
\description{
Animated placeholder for loading content. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
