% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sliders.R
\name{Slider}
\alias{Slider}
\alias{updateMantineSlider}
\title{Mantine Slider (Shiny stateful input)}
\usage{
Slider(inputId, value = 0, ...)

updateMantineSlider(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced by
dragging the handle.}

\item{value}{Initial value.}

\item{...}{Other props (\code{min}, \code{max}, \code{step}, \code{marks}, \code{label}, \code{color},
...). See \url{https://mantine.dev/core/slider/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine Slider (Shiny stateful input)
}
