% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Space}
\alias{Space}
\title{Mantine Space}
\usage{
Space(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/space/}.}
}
\description{
Empty horizontal/vertical spacer. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
