% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{Splitter}
\alias{Splitter}
\alias{SplitterPanel}
\title{Mantine Splitter (resizable side-by-side panes)}
\usage{
Splitter(...)

SplitterPanel(...)
}
\arguments{
\item{...}{Props and children (usually two or more \code{\link[=SplitterPanel]{SplitterPanel()}}).
See \url{https://mantine.dev/core/splitter/}.}
}
\description{
Mantine Splitter (resizable side-by-side panes)
}
