% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Spoiler}
\alias{Spoiler}
\title{Mantine Spoiler}
\usage{
Spoiler(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/spoiler/}.}
}
\description{
Truncated content with a "show more" toggle. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
