% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Stack}
\alias{Stack}
\title{Mantine Stack (vertical flex layout)}
\usage{
Stack(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/stack/}.}
}
\description{
Mantine Stack (vertical flex layout)
}
