% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Switch.R
\name{Switch}
\alias{Switch}
\alias{updateMantineSwitch}
\title{Mantine Switch (Shiny stateful boolean input)}
\usage{
Switch(inputId, label = NULL, value = FALSE, ...)

updateMantineSwitch(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (\code{TRUE}/\code{FALSE})
is synced on every toggle.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props (\code{description}, \code{color}, \code{size}, \code{disabled}, ...).
See \url{https://mantine.dev/core/switch/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine Switch (Shiny stateful boolean input)
}
