% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Table}
\alias{Table}
\alias{TableThead}
\alias{TableTbody}
\alias{TableTfoot}
\alias{TableTr}
\alias{TableTh}
\alias{TableTd}
\alias{TableCaption}
\alias{TableScrollContainer}
\title{Mantine Table family}
\usage{
Table(...)

TableThead(...)

TableTbody(...)

TableTfoot(...)

TableTr(...)

TableTh(...)

TableTd(...)

TableCaption(...)

TableScrollContainer(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/table/}.}
}
\description{
For a table with search/sort or selection, compose these elements by
hand (see \code{inst/examples/tables-app.R}); for a drag-and-drop reorderable
table use \code{\link[=SortableTable]{SortableTable()}}.
}
