% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Text}
\alias{Text}
\title{Mantine Text}
\usage{
Text(...)
}
\arguments{
\item{...}{Props and children forwarded to Mantine's \code{Text} component
(\code{size}, \code{fw}, \code{c}, \code{ta}, ...). See \url{https://mantine.dev/core/text/}.}
}
\description{
Mantine Text
}
