% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialInputs.R
\name{Textarea}
\alias{Textarea}
\title{Mantine Textarea (Shiny stateful input, multi-line)}
\usage{
Textarea(inputId, label = NULL, value = "", ...)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced on
every keystroke.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props (\code{autosize}, \code{minRows}, \code{maxRows}, ...). See
\url{https://mantine.dev/core/textarea/}.}
}
\description{
Mantine Textarea (Shiny stateful input, multi-line)
}
