% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ThemeIcon}
\alias{ThemeIcon}
\title{Mantine ThemeIcon (colored container for icons)}
\usage{
ThemeIcon(...)
}
\arguments{
\item{...}{Props and children (\code{color}, \code{variant}, \code{size}, \code{radius},
...). See \url{https://mantine.dev/core/theme-icon/}.}
}
\description{
Mantine ThemeIcon (colored container for icons)
}
