% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{TimeGrid}
\alias{TimeGrid}
\alias{updateMantineTimeGrid}
\title{Mantine TimeGrid (Shiny stateful input, grid of selectable time slots)}
\usage{
TimeGrid(inputId, data, value = NULL, ...)

updateMantineTimeGrid(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a \code{"HH:mm"}
string, or \code{NULL}) is synced on selection.}

\item{data}{A character vector of selectable times (e.g. \code{c("09:00", "09:30", "10:00")}).}

\item{value}{Initial value.}

\item{...}{Other props (\code{format}, \code{simpleGridProps}, ...). See
\url{https://mantine.dev/dates/time-grid/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine TimeGrid (Shiny stateful input, grid of selectable time slots)
}
