% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{TimeInput}
\alias{TimeInput}
\title{Mantine TimeInput (Shiny stateful input, native time field)}
\usage{
TimeInput(inputId, label = NULL, value = "", ...)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a \code{"HH:mm"}
string) is synced on every change.}

\item{label}{Field label.}

\item{value}{Initial value (a \code{"HH:mm"} string).}

\item{...}{Other props (\code{withSeconds}, \code{minTime}, \code{maxTime}, ...). See
\url{https://mantine.dev/dates/time-input/}.}
}
\description{
Mantine TimeInput (Shiny stateful input, native time field)
}
