% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{Timeline}
\alias{Timeline}
\alias{TimelineItem}
\title{Mantine Timeline}
\usage{
Timeline(...)

TimelineItem(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/timeline/}.}
}
\description{
Vertical sequence of events. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
