% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Title}
\alias{Title}
\title{Mantine Title (h1-h6 headings)}
\usage{
Title(...)
}
\arguments{
\item{...}{Props and children (\code{order} = 1..6, \code{size}, ...). See
\url{https://mantine.dev/core/title/}.}
}
\description{
Mantine Title (h1-h6 headings)
}
