% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputs2.R
\name{Tooltip}
\alias{Tooltip}
\title{Mantine Tooltip}
\usage{
Tooltip(...)
}
\arguments{
\item{...}{Props and a single child (\code{label}, \code{position}, \code{withArrow},
...). See \url{https://mantine.dev/core/tooltip/}.}
}
\description{
Mantine Tooltip
}
