% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{TooltipFloating}
\alias{TooltipFloating}
\title{Mantine TooltipFloating}
\usage{
TooltipFloating(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/tooltip/}.}
}
\description{
Tooltip that follows the cursor instead of anchoring to the target. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
