% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-components.R
\name{TooltipGroup}
\alias{TooltipGroup}
\title{Mantine TooltipGroup}
\usage{
TooltipGroup(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/tooltip/}.}
}
\description{
Shares open/close delay behaviour across a group of Tooltips. Supports \code{mantineId} for reactive updates via
\code{\link[=updateMantineProps]{updateMantineProps()}}.
}
