% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{Tree}
\alias{Tree}
\title{Mantine Tree (hierarchical data display)}
\usage{
Tree(inputId = NULL, data, ...)
}
\arguments{
\item{inputId}{If set, \code{input[[inputId]]} receives the clicked node's
\code{value}.}

\item{data}{Tree data: a list of \code{list(value=, label=, children=...)}.}

\item{...}{Other props (\code{expandOnClick}, \code{levelOffset}, ...). See
\url{https://mantine.dev/core/tree/}.}
}
\description{
Renders nested/expandable nodes from \code{data} (a list of
\code{list(value=, label=, children=list(...))}). If \code{inputId} is set,
clicking a node reports its \code{value} to \code{input[[inputId]]}.
}
\examples{
\dontrun{
Tree(
  inputId = "tree_click",
  data = list(
    list(value = "src", label = "src", children = list(
      list(value = "src/app.R", label = "app.R")
    )),
    list(value = "readme", label = "README.md")
  )
)
}
}
