% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedComponents.R
\name{TreeSelect}
\alias{TreeSelect}
\alias{updateMantineTreeSelect}
\title{Mantine TreeSelect (select input with hierarchical data)}
\usage{
TreeSelect(inputId, data, value = NULL, ...)

updateMantineTreeSelect(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} is synced to the
selected value(s).}

\item{data}{Tree data: a list of \code{list(value=, label=, children=...)}.}

\item{value}{Initial value: a string, or a character vector if
\code{multiple = TRUE}.}

\item{...}{Other props (\code{multiple}, \code{placeholder}, ...). See
\url{https://mantine.dev/core/tree-select/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Behaves like \code{\link[=Select]{Select()}}/\code{\link[=MultiSelect]{MultiSelect()}} but for tree-shaped \code{data}.
}
