% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{UnstyledButton}
\alias{UnstyledButton}
\title{Mantine UnstyledButton (button with no default styling, for building custom elements)}
\usage{
UnstyledButton(...)
}
\arguments{
\item{...}{Props and children. See \url{https://mantine.dev/core/unstyled-button/}.}
}
\description{
Mantine UnstyledButton (button with no default styling, for building custom elements)
}
