% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{YearPickerInput}
\alias{YearPickerInput}
\alias{updateMantineYearPickerInput}
\title{Mantine YearPickerInput (Shiny stateful input, year selection)}
\usage{
YearPickerInput(inputId, label = NULL, value = NULL, ...)

updateMantineYearPickerInput(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  ...
)
}
\arguments{
\item{inputId}{Id of the Shiny input; \code{input[[inputId]]} (a
\code{"YYYY-MM-DD"} string, first day of the year) is synced on selection.}

\item{label}{Field label.}

\item{value}{Initial value.}

\item{...}{Other props. See \url{https://mantine.dev/dates/year-picker-input/}.}

\item{session}{Session object passed to the Shiny server function.}
}
\description{
Mantine YearPickerInput (Shiny stateful input, year selection)
}
